refactor project and fix imports
