+++ test/fuso_test.go
@@ -1,4 +1,4 @@
-package filedownloader
+package test
 
 import (
 	"fmt"
@@ -8,12 +8,15 @@ import (
 	_ "sync"
 	"testing"
 	"time"
+
+	fd "github.com/sysgoblin/filedownloader/cmd"
+	ihttp "github.com/sysgoblin/filedownloader/internal/http"
 )
 
 // filedownloader test
 
 func TestSimpleSingleDownload(t *testing.T) {
-	fdl := New(nil)
+	fdl := fd.New(nil)
 	user, _ := user.Current()
 	err := fdl.SimpleFileDownload(`https://golang.org/pkg/net/http/`, user.HomeDir+`/fuso.html`)
 	if err != nil {
@@ -22,12 +25,12 @@ func TestSimpleSingleDownload(t *testing.T) {
 }
 
 func TestMultipleFilesDownload(t *testing.T) {
-	fdl := New(nil)
+	fdl := fd.New(nil)
 	user, _ := user.Current()
 	// Download Progress Observer
-	var downloadFiles []*Download
-	downloadFiles = append(downloadFiles, &Download{URL: `https://files.hareruyamtg.com/img/goods/L/M21/EN/0001.jpg`, LocalFilePath: user.HomeDir + `/ugin.jpg`})
-	downloadFiles = append(downloadFiles, &Download{URL: `https://files.hareruyamtg.com/img/goods/L/ELD/EN/BRAWL0329.jpg`, LocalFilePath: user.HomeDir + `/korvold.jpg`})
+	var downloadFiles []*fd.Download
+	downloadFiles = append(downloadFiles, &fd.Download{URL: `https://files.hareruyamtg.com/img/goods/L/M21/EN/0001.jpg`, LocalFilePath: user.HomeDir + `/ugin.jpg`})
+	downloadFiles = append(downloadFiles, &fd.Download{URL: `https://files.hareruyamtg.com/img/goods/L/ELD/EN/BRAWL0329.jpg`, LocalFilePath: user.HomeDir + `/korvold.jpg`})
 	err := fdl.MultipleFileDownload(downloadFiles)
 	if err != nil {
 		t.Error(err)
@@ -39,17 +42,17 @@ func TestFloatProgressCalc(t *testing.T) {
 	fmt.Println(v)
 }
 
-func TestExternalLogFunction(t *testing.T) {
-	conf := Config{logfunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3}
-	fileDownloader := New(&conf)
+func TestExternalLogFunc(t *testing.T) {
+	conf := fd.Config{LogFunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3}
+	fileDownloader := fd.New(&conf)
 	// downloading to use home
 	user, _ := user.Current()
 	fileDownloader.SimpleFileDownload(`https://golang.org/pkg/net/http/`, user.HomeDir+`/fuso.html`)
 }
 
 func TestCancelWhileDownloading(t *testing.T) {
-	conf := Config{logfunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3, MaxRetry: 3}
-	fileDownloader := New(&conf)
+	conf := fd.Config{LogFunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3, MaxRetry: 3}
+	fileDownloader := fd.New(&conf)
 	// downloading to use home
 	user, _ := user.Current()
 	go func() {
@@ -63,16 +66,16 @@ func TestCancelWhileDownloading(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	if fileDownloader.err != nil {
-		t.Error(fileDownloader.err)
+	if fileDownloader.Err != nil {
+		t.Error(fileDownloader.Err)
 	}
 	t.Log(`Test Done`)
 }
 
 func TestFileDownloadWithDetailedConfiguration(t *testing.T) {
 	// default setting of RequiresDetailProgress is false, you need to set it true if you need download progress.
-	conf := Config{logfunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3, MaxRetry: 3, RequiresDetailProgress: true}
-	fileDownloader := New(&conf)
+	conf := fd.Config{LogFunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3, MaxRetry: 3, RequiresDetailProgress: true}
+	fileDownloader := fd.New(&conf)
 
 	done := make(chan int)
 	// if you set RequiresDetailProgress = true, you can receive progress from channel
@@ -82,10 +85,10 @@ func TestFileDownloadWithDetailedConfiguration(t *testing.T) {
 			select {
 			case speed := <-fileDownloader.DownloadBytesPerSecond:
 				// DownloadBytesPerSecond Channel can receive how fast the download is running.
-				log.Println(fmt.Sprintf(`%d bytes/sec`, speed))
+				log.Printf(`%d bytes/sec`, speed)
 			case progress := <-fileDownloader.ProgressChan:
 				// Progress Channel (ProgressChan) receives how much download has progressed.
-				log.Println(fmt.Sprintf(`%f percent has done`, progress*100)) // ex. 10.5 percent has done
+				log.Printf(`%f percent has done`, progress*100) // ex. 10.5 percent has done
 			case <-done:
 				break LOOP // escape from forever loop
 			}
@@ -101,16 +104,16 @@ func TestFileDownloadWithDetailedConfiguration(t *testing.T) {
 		t.Error(err)
 		done <- 1
 	}
-	if fileDownloader.err != nil {
-		t.Error(fileDownloader.err)
+	if fileDownloader.Err != nil {
+		t.Error(fileDownloader.Err)
 	}
 	done <- 0
 	t.Log(`Test Done`)
 }
 
 func TestMultiFileDownloadCancelWhileDownloading(t *testing.T) {
-	conf := Config{logfunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3, MaxRetry: 3}
-	fileDownloader := New(&conf)
+	conf := fd.Config{LogFunc: myLogger, MaxDownloadThreads: 1, DownloadTimeoutMinutes: 3, MaxRetry: 3}
+	fileDownloader := fd.New(&conf)
 	// downloading to use home
 	user, _ := user.Current()
 	go func() {
@@ -119,16 +122,16 @@ func TestMultiFileDownloadCancelWhileDownloading(t *testing.T) {
 		// wait and cancel
 		fileDownloader.Cancel()
 	}()
-	var downloadFiles []*Download
-	downloadFiles = append(downloadFiles, &Download{URL: "http://ipv4.download.thinkbroadband.com/512MB.zip", LocalFilePath: user.HomeDir + `/512.zip`})
-	downloadFiles = append(downloadFiles, &Download{URL: "http://ipv4.download.thinkbroadband.com/200MB.zip", LocalFilePath: user.HomeDir + `/200.zip`})
+	var downloadFiles []*fd.Download
+	downloadFiles = append(downloadFiles, &fd.Download{URL: "http://ipv4.download.thinkbroadband.com/512MB.zip", LocalFilePath: user.HomeDir + `/512.zip`})
+	downloadFiles = append(downloadFiles, &fd.Download{URL: "http://ipv4.download.thinkbroadband.com/200MB.zip", LocalFilePath: user.HomeDir + `/200.zip`})
 	// test download file 512MB
 	err := fileDownloader.MultipleFileDownload(downloadFiles)
 	if err != nil {
 		t.Error(err)
 	}
-	if fileDownloader.err != nil {
-		t.Error(fileDownloader.err)
+	if fileDownloader.Err != nil {
+		t.Error(fileDownloader.Err)
 	}
 	t.Log(`Test Done`)
 }
@@ -139,7 +142,7 @@ func myLogger(params ...interface{}) {
 
 func TestFileExists(t *testing.T) {
 	user, _ := user.Current()
-	bytes, err := getFileStartOffset(user.HomeDir + `/512.zip`)
+	bytes, err := ihttp.GetFileStartOffset(user.HomeDir + `/512.zip`)
 	if err != nil {
 		t.Log(err)
 		return
